+++ Fotogroep Waalre/ViewModel/IndividualClubs/FotogroepAnders/AndersMembersProvider.swift
@@ -11,7 +11,7 @@ class AndersMembersProvider {
 
     init(bgContext: NSManagedObjectContext) {
 
-        // Photo club Bellus Imago doesn't currently support loading member data from an online site.
+        // Photogroup Anders doesn't currently support loading member data from an online site.
         // So we only insert a member or two from info hardcoded in insertSomeHardcodedMemberData()
         insertSomeHardcodedMemberData(bgContext: bgContext)
     }